Use CTest resource spec to distribute tests on parallel runs
hipCUB tests are often run in parallel when they go through CI
checks. Right now, on machines with more than a single GPU,
parallel tests are not evenly distributed across devices. This
can cause devices that are overloaded to run out of memory,
which leads to test failures or terminated subprocess errors.

This change adds a ctest "resource spec" file that contains
information about the available GPUs in the system. Using this
file, ctest can distribute tests evenly across multiple devices.
Specifically, this change:
- automatically generates a resource spec file at cmake build time.
Set CMake variables to cause this resource file to be used when
the ctest command is issued from the build directory.
- On install, regenerate the resource spec file within the install folder.
We regenerate this so that when hipCUB is installed via a package that
was generated with cpack, we don't end up using the resource file that
was generated when the package was built (since that may have happened
on a different system with different devices).
I've added code to the CTestTestfile that's generated for installation so
that the newly generated resource spec is picked up by ctest correctly within
the installation directory.

In theory, the above two changes should make the use of the resource spec
pretty much transparent to the user - so QA/CQE should not need to change
their build scripts. There is one caveat though:
  - If the test system has more than one kind of device, by default
    tests will be distributed across all devices.
  - If you want to limit the tests to running on only a single device, you
    will need to set -DGPU_TEST_TARGETS=<gfx ID> at cmake configure time.